sccp-receptores/requirements.txt
paho-mqtt==1.6.1
rpyc
